HW0-Installation/requirements.txt
numpy
scipy
scikit-image
matplotlib
ipython
jupyter
jupyterlab
ipykernel
pandas
sympy
nose
opencv-python
pillow
pytest
ipytest
glob2
rawpy
pathlib